Change LIGO authentication to work with a domain hop
Summary:
In the current implementation, if a user logs in with ligo details, but the application is expected to be running at a different domain, the /auth/ligo/ view handled by gw-cloud.org doesn't set auth tokens that can be consumed by the correct domain (gwcloud.org.au). This diff adds a new view which allows a JWT token to be generated by /auth/ligo/, and passed by redirect (this is where the webserver will return a 302 to the new domain), to /auth/ligo_continue/. At this point the request will be under the correct domain, and the token is then unpacked, and a script returned that sets the correct tokens in the browser and redirects the user to the expected endpoint.

The new flow looks like:
* User visits /auth/ligo/ (gw-cloud.org)
* Shibboleth redirects the user to the LIGO login.
* Ligo login redirets a logged in user to /auth/ligo/ (gw-cloud.org)
* /auth/ligo/ generates a JWT token with correct login tokens (gw-cloud.org)
* /auth/ligo/ returns a redirect to /auth/ligo_continue/ with the encrypted token as a get paramater (gw-cloud.org)
* Webserver catches the browser making the request to gw-cloud.org and redirects the request to gwcloud.org.ou
* User's browser makes the request to gwcloud.org.au /auth/ligo_continue with the token provided from /auth/ligo/ (gwcloud.org.au)
* /auth/ligo_continue decodes the JWT token, and returns a script which sets the localstorage tokens, as /auth/ligo/ used to (gwcloud.org.au)

Test Plan: Passes review. Works in production.

Reviewers: jesmigel, mr-asher

Reviewed By: jesmigel, mr-asher

Differential Revision: https://phab.adacs.org.au/D225