Add user account registration back in
Summary: Since we want to use the auth module with GWLab, we need to allow users to register and use the system with non-ligo accounts. This diff finishes porting the register/verify pages over to bootstrap and enables user registration again in the backend.

Test Plan: Tests pass, works in prod, passes review

Reviewers: mr-asher

Reviewed By: mr-asher

Differential Revision: https://phab.adacs.org.au/D290