Update design for Login, Register and verify email pages.
Summary: Update the layout and design to accommodate users authenticating with their GWCloud credentials or Ligo credentials. This includes adding a well designed login form with username and password, adding a registration form with minimal inputs for ease of use, and a clear verification message.

Test Plan: Users should now be able to register a new account and use it to sign up.

Reviewers: llakerink, TomReichardt

Reviewed By: llakerink

Differential Revision: https://phab.adacs.org.au/D299