Support log in for both gwlab and gwcloud
Summary:
Currently when you try to use ligo auth from gwlab, you'll end up being redirected and logged in to gwcloud.org.au. This diff adds support for passing through a new domain query parameter which instructs the ligo login process to correctly redirect to gwlab or gwcloud respectively.

Fixes: T173

Test Plan:
Unfortunately this is very difficult to test outside of production, and therefor requires close inspection of the changes in views.py and Login.js

The new code should clearly provide a mechanism to pass a new domain query parameter for ligo auth which correctly triggers a redirect to the relevant domain along with the JWT token to be unpacked in to local storage.

There are additional changes to fix missing next url for ligo login as well.

Reviewers: mr-asher, TomReichardt, jesmigel

Reviewed By: mr-asher

Differential Revision: https://phab.adacs.org.au/D347