Make sure newly registered users can't log in without verifying their account
Summary:
Currently when a new user registers, they can log in immediately rather than having to verify their account. This is not desired behavour. This diff makes sure than new user accounts are not marked as is_active=True.

Fixes: T174

After testing, some additional things were caught, including first and last name could be empty, and verification tokens could be reused.

Test Plan:
Using this patch locally means you should be able to create a new user account which is not yet verified. While the account is not verified, logging in with the user accounts details should be rejected. If you then verify the account using the url from the verification email, you should now be able to log in successfully with the new user account.

When registering, first and last name should no longer be able to be empty (via graphi), and verification tokens should no longer be able to be used more than once.

Reviewers: mr-asher

Reviewed By: mr-asher

Differential Revision: https://phab.adacs.org.au/D348