Update JWT payload to include a flag for if the user is ligo authenticated or not
Summary:
Until this point there has been no need to protect ligo proprietary data because we haven't been using it. Now, with the Viterbi module, we'll be using O3 (proprietary data) and it is important that only users who are affiliated with ligo can access this data. We're already storing details about if a user is a ligo user on not in the database, we just need a way to pass that flag through to applications for consumption.

This diff adds an extra JWT payload flag indicating if the user is authenticated with ligo or not.

Test Plan: Checking out this diff and running the tests should not fail.

Reviewers: mr-asher, jesmigel

Reviewed By: mr-asher

Differential Revision: https://phab.adacs.org.au/D371